refactor(linter/plugins): move comments-related code into separate file (#14753)
Follow-on after #14715. Pure refactor. Move code for `SourceCode`'s comment-related methods into a separate file. Code is completely unchanged - it's just moved.
